requirements-docker.txt
# Web Framework
uvicorn>=0.24.0
fastapi>=0.109.0
pydantic-settings>=2.0.0

# Database
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
alembic>=1.13.0

# Caching & Sessions
redis>=5.0.0
aioredis>=2.0.1

# Object Storage
minio>=7.2.0

# Observability & Tracing
prometheus-client>=0.19.0
opentelemetry-api>=1.21.0
opentelemetry-sdk>=1.21.0
opentelemetry-exporter-jaeger>=1.21.0
opentelemetry-instrumentation-fastapi>=0.42b0
opentelemetry-instrumentation-sqlalchemy>=0.42b0
opentelemetry-instrumentation-redis>=0.42b0
opentelemetry-instrumentation-psycopg2>=0.42b0

# Logging
python-json-logger>=2.0.0
structlog>=24.1.0

# Health checks
httpx>=0.25.0
